Move artifact types out of hardhat network.
