+++ packages/hardhat-core/src/internal/artifacts.ts
@@ -3,8 +3,7 @@ import fsExtra from "fs-extra";
 import * as os from "os";
 import * as path from "path";
 
-import { Artifact } from "../types/artifact";
-import { SolcInput } from "../types/solidity";
+import { Artifact, CompilerInput } from "../types";
 
 import { BUILD_INFO_DIR_NAME } from "./constants";
 import { assertHardhatInvariant, HardhatError } from "./core/errors";
@@ -226,7 +225,7 @@ export class Artifacts {
    * to the written file.
    */
   public async saveBuildInfo(
-    input: SolcInput,
+    input: CompilerInput,
     output: any,
     solcVersion: string
   ): Promise<string> {
