+++ packages/hardhat-core/src/internal/hardhat-network/provider/input.ts
@@ -2,7 +2,7 @@ import { BN, isValidAddress, toBuffer } from "ethereumjs-util";
 import * as t from "io-ts";
 import { PathReporter } from "io-ts/lib/PathReporter";
 
-import { CompilerInput, CompilerOutput } from "../stack-traces/compiler-types";
+import { CompilerInput, CompilerOutput } from "../../../types";
 
 import { InvalidArgumentsError } from "./errors";
