+++ packages/hardhat-core/src/internal/solidity/compiler/compiler-input.ts
@@ -1,11 +1,13 @@
-import { CompilationJob } from "../../../types/builtin-tasks";
-import { SolcOptimizerConfig } from "../../../types/config";
-import { SolcInput } from "../../../types/solidity";
+import {
+  CompilationJob,
+  CompilerInput,
+  SolcOptimizerConfig,
+} from "../../../types";
 import { DependencyGraph } from "../dependencyGraph";
 
 export function getInputFromCompilationJob(
   compilationJob: CompilationJob
-): SolcInput {
+): CompilerInput {
   const sources: { [sourceName: string]: { content: string } } = {};
   for (const file of compilationJob.getResolvedFiles()) {
     sources[file.sourceName] = {
@@ -15,7 +17,7 @@ export function getInputFromCompilationJob(
 
   const { settings } = compilationJob.getSolcConfig();
 
-  const input: SolcInput = {
+  const input: CompilerInput = {
     language: "Solidity",
     sources,
     settings: {
