+++ packages/hardhat-core/src/types/index.ts
@@ -3,5 +3,5 @@ export * from "./plugins";
 export * from "./experimental";
 export * from "./network";
 export * from "./runtime";
-export * from "./solidity";
 export * from "./artifact";
+export * from "./builtin-tasks";
