--- packages/hardhat-core/src/types/solidity.ts
@@ -1,12 +1,0 @@
-import { SolcOptimizerConfig } from "./config";
-
-export interface SolcInput {
-  settings: {
-    metadata: { useLiteralContent: boolean };
-    optimizer: SolcOptimizerConfig;
-    outputSelection: { "*": { "": string[]; "*": string[] } };
-    evmVersion?: string;
-  };
-  sources: { [p: string]: { content: string } };
-  language: string;
-}
